Add getObject to IdeConnection instead of just IdeSession
- Need some way to get objects when sessions are unavailable
